Refactor common code from VDF, CIF and IRF into a common base class.
